+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
         android:theme="@style/Theme.ShoppingCart"
         tools:targetApi="31">
         <activity
-            android:name=".CartActivity"
+            android:name=".features.cart.CartActivity"
             android:exported="false" />
         <activity
-            android:name=".ProductDetailActivity"
+            android:name=".features.productdetail.ProductDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".features.productlist.ProductListActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ShoppingCart">
