+++ app/src/main/java/nextstep/shoppingcart/components/topbars/CenterTitleTopBar.kt
@@ -1,11 +1,14 @@
-package nextstep.shoppingcart.components.topbar
+@file:OptIn(ExperimentalMaterial3Api::class)
 
+package nextstep.shoppingcart.components.topbars
+
+import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.TopAppBarScrollBehavior
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -13,39 +16,43 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-internal fun StartTitleTopBar(
+@OptIn(ExperimentalMaterial3Api::class)
+internal fun CenterTitleTopBar(
     title: String,
-    navigationType: TopBarNavigationType,
-    onNavigationClick: () -> Unit,
+    action: TopBarActionType,
+    onActionClick: () -> Unit,
     modifier: Modifier = Modifier,
+    scrollBehavior: TopAppBarScrollBehavior? = null,
 ) {
-    TopAppBar(
+    CenterAlignedTopAppBar(
         title = {
             Text(text = title)
         },
-        navigationIcon = {
-            IconButton(onClick = onNavigationClick) {
+        actions = {
+            IconButton(
+                onClick = onActionClick,
+            ) {
                 Icon(
-                    imageVector = navigationType.icon,
-                    contentDescription = stringResource(navigationType.contentDescription),
+                    imageVector = action.icon,
+                    contentDescription = stringResource(action.contentDescription),
                 )
             }
         },
         colors = TopAppBarDefaults.topAppBarColors().copy(containerColor = Color.White),
+        scrollBehavior = scrollBehavior,
         modifier = modifier,
     )
 }
 
 @Preview
 @Composable
-private fun StartTitleTopBarPreview() {
+private fun CenterTitleTopBarPreview() {
     ShoppingCartTheme {
-        StartTitleTopBar(
-            title = "장바구니",
-            navigationType = TopBarNavigationType.BACK,
-            onNavigationClick = {},
+        CenterTitleTopBar(
+            title = "상품 목록",
+            action = TopBarActionType.CART,
+            onActionClick = {},
         )
     }
 }
