+++ app/src/main/java/nextstep/shoppingcart/components/topbars/StartTitleTopBar.kt
@@ -1,14 +1,11 @@
-@file:OptIn(ExperimentalMaterial3Api::class)
+package nextstep.shoppingcart.components.topbars
 
-package nextstep.shoppingcart.components.topbar
-
-import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.TopAppBarScrollBehavior
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -16,43 +13,39 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
-@Composable
 @OptIn(ExperimentalMaterial3Api::class)
-internal fun CenterTitleTopBar(
+@Composable
+internal fun StartTitleTopBar(
     title: String,
-    action: TopBarActionType,
-    onActionClick: () -> Unit,
+    navigationType: TopBarNavigationType,
+    onNavigationClick: () -> Unit,
     modifier: Modifier = Modifier,
-    scrollBehavior: TopAppBarScrollBehavior? = null,
 ) {
-    CenterAlignedTopAppBar(
+    TopAppBar(
         title = {
             Text(text = title)
         },
-        actions = {
-            IconButton(
-                onClick = onActionClick,
-            ) {
+        navigationIcon = {
+            IconButton(onClick = onNavigationClick) {
                 Icon(
-                    imageVector = action.icon,
-                    contentDescription = stringResource(action.contentDescription),
+                    imageVector = navigationType.icon,
+                    contentDescription = stringResource(navigationType.contentDescription),
                 )
             }
         },
         colors = TopAppBarDefaults.topAppBarColors().copy(containerColor = Color.White),
-        scrollBehavior = scrollBehavior,
         modifier = modifier,
     )
 }
 
 @Preview
 @Composable
-private fun CenterTitleTopBarPreview() {
+private fun StartTitleTopBarPreview() {
     ShoppingCartTheme {
-        CenterTitleTopBar(
-            title = "상품 목록",
-            action = TopBarActionType.CART,
-            onActionClick = {},
+        StartTitleTopBar(
+            title = "장바구니",
+            navigationType = TopBarNavigationType.BACK,
+            onNavigationClick = {},
         )
     }
 }
