+++ app/src/main/java/nextstep/shoppingcart/features/cart/CartActivity.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart
+package nextstep.shoppingcart.features.cart
 
 import android.content.Context
 import android.content.Intent
@@ -7,21 +7,22 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import nextstep.shoppingcart.data.InMemoryCartRepository
-import nextstep.shoppingcart.screens.cart.CartScreen
+import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 class CartActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             enableEdgeToEdge()
-
-            CartScreen(
-                cart = InMemoryCartRepository.cart.value,
-                onBackClick = onBackPressedDispatcher::onBackPressed,
-                onAddOneClick = InMemoryCartRepository::addOne,
-                onRemoveOneClick = InMemoryCartRepository::removeOne,
-                onRemoveAllClick = InMemoryCartRepository::removeAll,
-            )
+            ShoppingCartTheme {
+                CartScreen(
+                    cart = InMemoryCartRepository.cart.value,
+                    onBackClick = onBackPressedDispatcher::onBackPressed,
+                    onAddOneClick = InMemoryCartRepository::addOne,
+                    onRemoveOneClick = InMemoryCartRepository::removeOne,
+                    onRemoveAllClick = InMemoryCartRepository::removeAll,
+                )
+            }
         }
     }
