+++ app/src/main/java/nextstep/shoppingcart/features/cart/CartScreen.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.screens.cart
+package nextstep.shoppingcart.features.cart
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,11 +7,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateMapOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
@@ -19,13 +14,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import nextstep.shoppingcart.R
 import nextstep.shoppingcart.components.MainButton
-import nextstep.shoppingcart.components.topbar.StartTitleTopBar
-import nextstep.shoppingcart.components.topbar.TopBarNavigationType
+import nextstep.shoppingcart.components.topbars.StartTitleTopBar
+import nextstep.shoppingcart.components.topbars.TopBarNavigationType
 import nextstep.shoppingcart.data.FakeProductRepository
 import nextstep.shoppingcart.domain.model.Cart
 import nextstep.shoppingcart.domain.model.CartItem
 import nextstep.shoppingcart.domain.model.Product
-import nextstep.shoppingcart.screens.cart.component.CartList
+import nextstep.shoppingcart.features.cart.component.CartList
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 @Composable
