+++ app/src/main/java/nextstep/shoppingcart/features/productlist/ProductListScreen.kt
@@ -1,4 +1,6 @@
-package nextstep.shoppingcart.screens
+@file:JvmName("ProductListScreenKt")
+
+package nextstep.shoppingcart.features.productlist
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -7,6 +9,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -17,13 +20,18 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
@@ -32,42 +40,60 @@ import androidx.compose.ui.unit.dp
 import nextstep.shoppingcart.R
 import nextstep.shoppingcart.components.ItemCounter
 import nextstep.shoppingcart.components.ProductImage
+import nextstep.shoppingcart.components.topbars.CenterTitleTopBar
+import nextstep.shoppingcart.components.topbars.TopBarActionType
 import nextstep.shoppingcart.data.FakeProductRepository
 import nextstep.shoppingcart.domain.model.Cart
 import nextstep.shoppingcart.domain.model.Product
 import nextstep.shoppingcart.domain.model.Products
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 internal fun ProductListScreen(
     products: Products,
     cart: Cart,
     onAddOneClick: (Product) -> Unit,
     onRemoveOneClick: (Product) -> Unit,
+    onActionCartClick: () -> Unit,
     onProductClick: (Product) -> Unit,
-    modifier: Modifier = Modifier,
 ) {
-    LazyVerticalGrid(
-        modifier = modifier.padding(horizontal = 18.dp),
-        columns = GridCells.Fixed(2),
-        verticalArrangement = Arrangement.spacedBy(20.dp),
-        horizontalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        items(
-            key = { it.id },
-            items = products.value
-        ) { product ->
-            Product(
-                product = product,
-                count = cart.find(product)?.count?.value,
-                onAddOneClick = { onAddOneClick(product) },
-                onRemoveOneClick = { onRemoveOneClick(product) },
-                onProductClick = { onProductClick(product) },
+    val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior(rememberTopAppBarState())
+    Scaffold(
+        topBar = {
+            CenterTitleTopBar(
+                title = stringResource(R.string.product_list_top_bar_title),
+                action = TopBarActionType.CART,
+                onActionClick = onActionCartClick,
+                scrollBehavior = scrollBehavior,
             )
+        },
+        modifier = Modifier
+            .fillMaxSize()
+            .nestedScroll(scrollBehavior.nestedScrollConnection),
+        containerColor = Color.White,
+    ) { paddingValues ->
+        LazyVerticalGrid(
+            modifier = Modifier.padding(paddingValues).padding(horizontal = 18.dp),
+            columns = GridCells.Fixed(2),
+            verticalArrangement = Arrangement.spacedBy(20.dp),
+            horizontalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            items(
+                key = { it.id },
+                items = products.value
+            ) { product ->
+                Product(
+                    product = product,
+                    count = cart.find(product)?.count?.value,
+                    onAddOneClick = { onAddOneClick(product) },
+                    onRemoveOneClick = { onRemoveOneClick(product) },
+                    onProductClick = { onProductClick(product) },
+                )
+            }
         }
     }
 }
-
 @Composable
 internal fun Product(
     product: Product,
@@ -148,13 +174,25 @@ private fun ProductImageAndCounter(
 @Composable
 private fun ProductListScreenPreview() {
     ShoppingCartTheme {
-        ProductListScreen(
-            products = FakeProductRepository.getAllProducts(),
-            cart = Cart(),
-            onAddOneClick = {},
-            onRemoveOneClick = {},
-            onProductClick = {},
-        )
+        LazyVerticalGrid(
+            modifier = Modifier.padding(horizontal = 18.dp),
+            columns = GridCells.Fixed(2),
+            verticalArrangement = Arrangement.spacedBy(20.dp),
+            horizontalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            items(
+                key = { it.id },
+                items = FakeProductRepository.getAllProducts().value
+            ) { product ->
+                Product(
+                    product = product,
+                    count = Cart().find(product)?.count?.value,
+                    onAddOneClick = { },
+                    onRemoveOneClick = { },
+                    onProductClick = { },
+                )
+            }
+        }
     }
 }
